Define unit(::Any)m unit(::Funciton) and add a type restriction to unit(::Type{Union{Missing,T}})
This is to resolve errors with uncluding units into FiniteDifferences.jl.
See https://github.com/JuliaDiff/FiniteDifferences.jl/pull/244 for details.

Tests for Units.jl need to be updated, if this is to be accepted.
The totalness of unit introduced via unit(::Ay) might be a bit controversial,
but I don't think that should be a problem. In general, things don't have units.
Only Numbers do.

Closes #806.